8. Git подключен через Retrofit.
