+++ app/src/main/java/ru/ds/gitapp/model/MockProjectsRepoImpl.kt
@@ -1,11 +1,9 @@
 package ru.ds.gitapp.model
 
-import io.reactivex.rxjava3.core.Single
-import ru.ds.gitapp.domain.GitUserRep
-
+/*
 class MockProjectsRepoImpl : GitUserRep {
 
-    override fun getUsers(userName: String): Single<List<GitUserEntity>> {
+    override fun getUsers(username: String): Single<List<GitUserEntity>> {
         val mockList = listOf(
             GitUserEntity("0", "!!!"),
             GitUserEntity("0", "!!!"),
@@ -16,4 +14,4 @@ class MockProjectsRepoImpl : GitUserRep {
         return Single.just(mockList)
 
     }
-}
+}*/
