+++ app/src/main/java/ru/ds/gitapp/model/retrofit/RetrofitUsersRepoImpl.kt
@@ -1,0 +1,38 @@
+package ru.ds.gitapp.model.retrofit
+
+import io.reactivex.rxjava3.core.Single
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+import ru.ds.gitapp.domain.GitUserRep
+import ru.ds.gitapp.model.GitUserEntity
+
+//Именно эту имплементацию (retrofit) в App
+class RetrofitUsersRepoImpl : GitUserRep {
+    private val retrofit = Retrofit.Builder()
+        .baseUrl("https://api.github.com/")
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+    private val api: GitHubApi = retrofit.create(GitHubApi::class.java)
+
+    override fun getUsers(username: String): Single<List<GitUserEntity>> {
+        return Single.create { emitter ->
+            api.listRepos(username)
+                .enqueue(object : Callback<List<GitUserEntity>> {
+                    override fun onResponse(
+                        call: Call<List<GitUserEntity>>,
+                        response: Response<List<GitUserEntity>>
+                    ) {
+                        emitter.onSuccess(response.body())
+                    }
+
+                    override fun onFailure(call: Call<List<GitUserEntity>>, t: Throwable) {
+                        emitter.onError(t)
+                    }
+                })
+        }
+    }
+
+}
