+++ app/src/main/java/ru/ds/gitapp/ui/gitusers/GitAdapter.kt
@@ -1,4 +1,4 @@
-package ru.ds.gitapp.ui.about
+package ru.ds.gitapp.ui.gitusers
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
