+++ app/src/main/java/ru/ds/gitapp/ui/gitusers/GitUsersFragment.kt
@@ -5,14 +5,20 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ru.ds.gitapp.databinding.SettingsFragmentBinding
+import androidx.fragment.app.viewModels
+import androidx.recyclerview.widget.LinearLayoutManager
+import ru.ds.gitapp.app
+import ru.ds.gitapp.databinding.AboutFragmentBinding
+import ru.ds.gitapp.viewModel.GITRepositoryViewModel
 
 
-
-class GitUsersFragment : Fragment() {
-    private var _binding: SettingsFragmentBinding? = null
-    private val binding: SettingsFragmentBinding
+class AboutFragment : Fragment() {
+    private var _binding: AboutFragmentBinding? = null
+    private val binding: AboutFragmentBinding
         get() = _binding!!
+
+    private val viewModel: GITRepositoryViewModel by viewModels { ReposViewModelFactory(requireActivity().app.gitUserRepo) }
+    private val adapter = GitAdapter()
 
 
     override fun onCreateView(
@@ -20,9 +26,40 @@ class GitUsersFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = SettingsFragmentBinding.inflate(inflater, container, false)
+        _binding = AboutFragmentBinding.inflate(inflater, container, false)
         return binding.root
     }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initOutgoingEvents() //метод отправляет event
+        initIncomingEvents() //метод получает event
+        recyclerView() //отображение данных в recycler View
+
+    }
+
+    private fun recyclerView() {
+        binding.recycleView.layoutManager = LinearLayoutManager(requireContext())
+        adapter.setHasStableIds(true)
+        binding.recycleView.adapter = adapter
+
+    }
+
+
+    private fun initIncomingEvents() {
+        binding.buttonEnterText.setOnClickListener {
+            val userData = binding.enterEditText.text.toString()
+            viewModel.onShowRepository(userData)
+        }
+    }
+
+    private fun initOutgoingEvents() {
+        viewModel.repo.observe(requireActivity()){
+           adapter.setData(it)
+        }
+    }
+
     override fun onDestroy() {
         super.onDestroy()
         _binding = null
@@ -30,6 +67,6 @@ class GitUsersFragment : Fragment() {
 
     companion object {
         @JvmStatic
-        fun newInstance() = GitUsersFragment()
+        fun newInstance() = AboutFragment()
     }
 }
