+++ app/src/main/res/layout/about_fragment.xml
@@ -13,7 +13,8 @@
             android:id="@+id/enter_edit_text"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="4" />
+            android:layout_weight="4"
+            android:text="@string/any"/>
 
         <Button
             android:id="@+id/button_enter_text"
