+++ app/src/main/res/values/strings.xml
@@ -9,4 +9,5 @@
     <string name="Email">EMAIL</string>
     <string name="profile">Local Users</string>
     <string name="User_name">ENTER</string>
+    <string name="any">name</string>
 </resources>
